requirements.txt
discord.py==2.5.2
PyTumblr==0.1.2
Django>=5.2.3,<6.0.0
django-extensions==4.1
Markdown==3.8.0
Pillow>=11.2.1,<11.3.0
psycopg2==2.9.10
#python-memcached==1.59
pymemcache==4.0.0
requests==2.32.4
natsort==8.4.0
atproto==0.0.61
